requirements.txt
# Core dependencies
pyyaml>=6.0
pdfplumber>=0.10.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
